Move children of Node (except Workflow) into submodule
and expose a dev API
