+++ pyiron_workflow/nodes/composite.py
