+++ pyiron_workflow/nodes/for_loop.py
@@ -11,9 +11,9 @@ from pandas import DataFrame
 from pyiron_snippets.factory import classfactory
 
 from pyiron_workflow.channels import NOT_DATA
-from pyiron_workflow.composite import Composite
+from pyiron_workflow.nodes.composite import Composite
 from pyiron_workflow.io_preview import StaticNode
-from pyiron_workflow.transform import inputs_to_dict, inputs_to_dataframe, InputsToDict
+from pyiron_workflow.nodes.transform import inputs_to_dict, inputs_to_dataframe, InputsToDict
 
 
 def dictionary_to_index_maps(
