+++ pyiron_workflow/nodes/function.py
@@ -45,7 +45,7 @@ class Function(StaticNode, ScrapesIO, ABC):
         At the most basic level, to use nodes all we need to do is provide the
         `Function` class with a function and labels for its output, like so:
 
-        >>> from pyiron_workflow.function import function_node
+        >>> from pyiron_workflow import function_node
         >>>
         >>> def mwe(x, y):
         ...     return x+1, y-1
@@ -195,7 +195,7 @@ class Function(StaticNode, ScrapesIO, ABC):
         for the return values, :param:output_labels, which are otherwise scraped from
         the text of the function definition:
 
-        >>> from pyiron_workflow.function import as_function_node
+        >>> from pyiron_workflow import as_function_node
         >>>
         >>> @as_function_node("p1", "m1")
         ... def my_mwe_node(
@@ -217,7 +217,7 @@ class Function(StaticNode, ScrapesIO, ABC):
         already defined as a `staticmethod`:
 
         >>> from typing import Literal, Optional
-        >>> from pyiron_workflow.function import Function
+        >>> from pyiron_workflow import Function
         >>>
         >>> class AlphabetModThree(Function):
         ...
