+++ tests/integration/test_transform.py
@@ -1,7 +1,7 @@
 import pickle
 import unittest
 
-from pyiron_workflow.transform import (
+from pyiron_workflow.nodes.transform import (
     inputs_to_list,
     inputs_to_list_factory,
     list_to_outputs,
