+++ tests/integration/test_workflow.py
@@ -1,13 +1,12 @@
 import math
 import pickle
 import random
-import time
 import unittest
 
 from pyiron_workflow._tests import ensure_tests_in_python_path
 from pyiron_workflow.channels import OutputSignal
-from pyiron_workflow.function import Function
-from pyiron_workflow.while_loop import while_loop
+from pyiron_workflow.nodes.function import Function
+from pyiron_workflow.nodes.while_loop import while_loop
 from pyiron_workflow.workflow import Workflow
